requirements.txt
streamlit
torch
torchvision
pillow
numpy
pandas
python-dotenv
requests
langchain-google-genai
